app_default_25_11_13_00_41_48/requirements.txt
fastapi==0.95.1
uvicorn[standard]==0.22.0
sqlalchemy==2.0.19
asyncpg==0.27.0
alembic==1.11.1
pydantic==1.10.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.0
pytest-asyncio==0.21.0
discord.py==2.4.0
aioredis==2.0.1
loguru==0.7.0
starlette-rate-limit==0.1.5
babel==2.12.1
python-dotenv==1.0.0
